Extend --format to support a full format specifier
Previously, it just supported versions and then we added extra options
as needed. This change allows it to specify the version plus some
additional properties. The idea is that we will be able to specify the
same properties in the signature file itself so we only need to add
properties in one place and they can be specified both on the command
line and in the signature file.

Bug: 295577788
Test: ./gradlew
Change-Id: I177c05521e3539ee548a091d485744b5404cccd4