Support preventing unexpected access to options
There is an ongoing process of migrating code away from the monolithic
global `options` property to using more modular option groups that are
passed to where they are needed. One problem with that work is that it
is often difficult to determine whether a specific piece of code uses
`options` indirectly.

This change adds an `OptionsDelegate` for use by `options` instead of
just using a simple property backed by a Java global variable. That
allows code that does not want to use `options` to make it inaccessible
to any code that it calls, simply by calling
`OptionsDelegate.disallowAccess()`. From that point on, up until the
`options` is assigned a new value, any attempt to access it will result
in an error which will provide information about why access was
disallowed.

The intent is that code that is being cleaned of `options` accesses
will call `OptionsDelegate.disallowAccess()` which should lead to test
failures and then once those failures are fixed then the call can
remain to prevent any additional `options` accesses being added by
mistake.

In future it is expected that this behavior will flip and `options`
will be inaccessible by default and any code that wants to access
`options` will have to make sure that it is set first.

This change uses this to protect `MergeSignaturesCommand` and
`UpdateSignatureHeaderCommand` from indirectly depending on `options`.

Bug: 286023667
Test: ./gradlew
Change-Id: Id70373d1b4465da012a5ffe67206a1eb201531fb