Add `--add-nonessential-overrides-classes` flag
`add-additional-overrides` only adds overriding methods that are
essential to compile the stubs and its reverse dependencies. However,
this may not be enough for passing all build time checks, such as
hiddenapi or errorprone. Some classes that require additional overrides
to be speficied in the signature files can be passed as inputs via the
`--add-nonessential-overrides-classes` flag.
The flag enables all visible abstract and concrete overriding methods
to be shown in the signature files for the specified classes.

Test: ./gradlew
Bug: 298063727
Change-Id: I05c376b03b850cd86e927c31245e3deebb2f09bc