Add `migrating` file format specifier property
The addition of support for file format specifiers in the signature
files broke Android build as it uses
`--api-overloaded-method-order=source` to override the default order of
`signature`. That was merged into the `FileFormat` retrieved from the
`--use-same-format-as` file and then instead of generating a signature
file header like this:
```
  // Signature format: 2.0
```

It generated one like this:
```
  // Signature format: 2.0:overloaded-method-order=source
```

Which broke `m checkapi`. It should not have written out the property
into the signature file because that is only intended for use when
migrating signature files (without changing version) to fix bugs.

This change avoids that by adding the `migrating` property to make it
explicit when migrating. It behaves as follows:
* It allows a developer to provide an explanation as to why the files
  are migrating, e.g. by proving a link to a bug with more details.
* If it is not present then additional properties are not written out
  into the header.
* If additional properties are provided and `migrating` is allow then
  `migrating` must be provided. That is because no existing version
  supports properties for anything other than migration.
* It is only allowed in signature files and in the `--format` option
  in `update-signature-header`. It is not allow in the `--format`
  option on other commands.

Bug: 295577788
Test: ./gradlew
Change-Id: I76a32075acf684b716eac8849cad1fe200f26064