Simplify parsing of text class declarations
Previously, a class declaration was parsed as a TextTypeItem to get the name without type parameters, and parsing of type parameters was done separately. This now gets the name and parses type parameters in one step without needing to go through TextTypeItem, and also consolidates two methods (asTypeInfo and toType) on TextClassItem that had the same purpose.

Bug: 297575681
Test: all existing tests pass

Change-Id: I6d280fba9acc9dd1cc87cc010a63f2b0e8c7d1e7