Support using 5.0 and properties in --format
Previously, using something like `--format 5.0:kotlin-style-nulls=no`
on the command line would fail because it is disallowed in the
signature file format (as `5.0` supports properties) and they both use
`parseSpecifier` to parse it. However, it should be allowed on the
command line so that the properties can be customized.

This change adds a `fromSignatureFile` parameter to `parseSpecifier` so
that it is allowed on the command line but not from the signature file.

Bug: 295577788
Test: ./gradlew
Change-Id: I9a79fdb13e87f15c734d0e0d391d6863c706728f