UnflaggedApi: Ignore APIs that are elided
APIs that have an exactly identical signature to one in (referencable)
super class are not emited into signature files. Consequently, we do not
know if they are pre-existing or not.

This is somewhat critical for the UnflaggedApi lint, which only applies to
unfinalized API (vs regular lints which should be followed in general, but
for finalized API often cannot be); this CL therefore disables that lint
for elided APIs.

Bug: 297362755
Test: ./gradlew :test --tests ApiLintTest.*FlaggedApi*
Test: make checkapi
Change-Id: Iace800c24f6de73c34b30f70c309704c7618bd6c