Remove asTypeParameter from TypeItem
It can be replaced with the property on VariableTypeItem. Previously, the `context` parameter was needed for text codebases as `TextTypeItem`s did not save a reference to which type parameter they referred to when created, but they now do. This means `context` is not necessary, as it was known at the time the type item was created.

Bug: 301285165
Test: existing tests pass
Change-Id: I82c9647a132e89e5e30d4d4985ed8435beca9c23