Add regression tests for nested classes in non-emitted outer classes
Signature files can contain inner classes without a corresponding
outer class.
This happens in particular in the platform with system APIs,
because the signature files elide classes whose system API surface
is entirely equal to the public API surface.

Notable examples are android.Manifest.permission and the
android.R.<resource type> classes, because their containing classes
are otherwise empty.

ComparisionVisitor treats outer classes that are only implied by
inner classes as if they were absent. Therefore the outer class
is considered added (because it does exist as a real class in the
from source codebases). In this case, it does however still recurse
into added outer classes.

There are two issues with this:
- ApiLint runs on anything (and its children) that's added. It
  therefore visits the inner class, even if it is pre-existing.
- ComparisonVisitor has a subtle bug when preferring real over
  implied items when using MergedCodebase, resulting in the
  real children of a real item in the lower-priority codebase
  erroneously getting preference over real children of a
  implied item in the first codebase, breaking CompatibilityCheck

Bug: 299675771
Test: ./gradlew test
Change-Id: I2ea8d86eadd9e07adf493371ca773a5573b352a8