Add ApiPredicate.Config to loadFromJarFile
This change adds the parameter, passes it through to the `ApiPredicate`
constructor and then passes in a default instance from
`ConvertJarsToSignatureFiles`. That allows for the removal of the
`try...catch...` block in the default parameter value for the
`ApiPredicate` constructor's `config` parameter as accessing `options`
from here does not cause test failures.

Bug: 299280167
Test: ./gradlew
Change-Id: I9b15810228acb8f6be8b9098fbfd712772227642