ComparisonVisitor: add regression test for nested classes in stubs
Adds a regression test for an interaction between MergedCodebase
and classes nested in implied classes. This happens in the platform
with nested @SystemApi classes, where the outer class - if it contains
only public APIs - is elided from the signature file.

When removing duplicates, the ComparisonVisitor prefers real classes
over implied classes, but still merges their children together.
However, it then prefers the real containing class' children over
the implied containing class' children, even when the children are
all real, regardless of the priority of the codebases.

Bug: 299675771
Test: ./gradlew check --tests ComparisonVisitorTest
Change-Id: I3c39533f6767b9e10c40d574ecff19c7d0bb0f81