Stop FileFormat defaults being written to header and specifier
Previously, the `CustomizableProperty` implementations for the
defaultable properties returned the effective value of the property,
which included any defaults, and not the specified value of the
property. That resulted in the defaults being written out to the
header.

This change switches them to use the specified value (which may be
`null`) and changes the `stringFromFormat()` method to return `String?`
to allow an unspecified value to be returned as `null`.

Bug: 300052204
Test: ./gradlew
Change-Id: Ia1959ddbb01e2f6c3cc36cfce62ab1c337b032d1