Move top-level src directory into sub-project
This makes a copy of `version.properties` in the root project and the
new `metalava` sub-project copies that into its resources. That allows
the file to exist in a well known location independent of where it is
needed at runtime.

This leaves a copy of it in the `src/main/resources` directory as
removing it would break downstream code. However, once this has been
submitted the downstream code should switch to using the copy in the
root project at which point the copy can be removed.

Bug: 300224510
Test: ./gradlew
Change-Id: I65a975b503fcf9d4cf3f644380328bb05a08ad17