Simplify reporting of invalid property error
Previously, when an invalid property error was reported the
`createUsageErrorMessage` would use some hacks to output the usage line
and allowable options without the intervening command description. The
main reason it did that was to maintain backwards compatibility with
the previous behavior before Clikt was used. However, it really does
not matter if the command description is output and it certainly is not
worth hacking the code to do that.

So, this change simply outputs the full command help, including
description, removing the need for the hacks. It also takes the
opportunity to rename the method to something more meaningful.

Bug: 286023667
Test: ./gradlew
Change-Id: If49d09e5ef3b8c0b8b50ec8bf61ea00cdc5fa966