Add option to hide non-released SDK extensions
There is a window of time between when a new SDK extension is finalized,
and when it is released externally. SDK artifacts (api-versions.xml,
docs) published during this window should not refer to the new SDK
extension.

Add a new option to metalava to ignore any SDK extensions of version X
or greater.

A separate CL will teach the build system to pass in the new option.

Bug: 300592449
Test: ./gradlew
Change-Id: I6a9b5700a4691a584b57ad9251add7a8cd3deac1