Pass single input TestFile to ModelSuiteRunner
Previously, the `ModelSuiteRunner` was expected to check to make sure
that it only process the `signature` or `source` parameter if it knew
that it supported them. However, the preceding change will only run a
test on a specific `InputFormat` if the runner supports that so there
is no need to the runner to check and so there is no need to pass the
different formats as different parameters.

This changes removes the `signature` and `source` parameters from
`ModelSuiteRunner.createCodebaseAndRun()` and instead passes in a
single `TestFile`.

Bug: 294556178
Test: ./gradlew
Change-Id: I1987b16e4a1ce2f7b20f3baecebc678ec47c9305