Add ReporterEnvironment abstraction
Tests need to intercept the reports that are written, before they are
printed as lines of text, in order to perform additional checks on them
that would be tricky if they just received a block of text. This change
encapsulates the support for that into a `ReporterEnvironment`. In the
process it was noticed that the `rootFolder` field is never `null` so
the type of that was changed from `File?` to `File`.

Bug: 295013652
Test: ./gradlew
Change-Id: I479ceb294b67563e8976e4ff0bfd4ec1f6b9a1bc