Dedup BaseModelTest test runner methods
Previously, there were three separate test run methods, each of which
could only handle a specific input file, i.e.
* `runJavaAndKotlinTest` could handle java and kotlin but not signature
  files.
* `runCodebaseTest` could handle java and signature but not kotlin
   files.
* `runSourceCodebaseTest` could handle java but not kotlin.

This change dedups the latter two so that they can all take a varargs
list of `TestFile` objects and uses their extension to determine which
`InputSet` they belong to and so which providers they should be tested
on.

The `runJavaAndKotlinTest` is removed and uses are replaced with
`runCodebaseTest`. That was chosen instead of `runSourceCodebaseTest`
because the tests are not `SourceCodebase` specific and it will be
necessary to add test for signature files too.

Bug: 294556178
Test: ./gradlew
Change-Id: Icd83608989ef08edeff719aa9724f768500fb6f8