Replace defaultCommandFactory with optional defaultCommandName
Previously, a `defaultCommandFactory` was passed to `MetalavaCommand`
and it was responsible for creating and registering the default
command. This replaces that with a `defaultCommandName` which is
optional and if it is provided then must match one of the registered
sub-commands.

This avoids `MetalavaCommand` having to know how to create the default
command, and by making it optional allows the removal of some fake test
classes.

Bug: 286023667
Test: ./gradlew
Change-Id: I16cf161c8a829332dc5e09f5aca2c692a6d7e349