Add support for RequiresFeature.enforcement
Previously, the `RequiredFeatures.enforcement` attribute was ignored
when generating the JavaDoc stubs and it always included text that said
to use the `PackageManager.hasSystemFeature(...)` method.

This change reads the `enforcement` attribute, makes sure that it
matches the expected pattern and then uses that to generate the text.
If the `enforcement` attribute does not match the pattern then a lint
issue is reported and it uses the default.

Bug: 288570779
Test: ./gradlew
Change-Id: Id0834c20370bd4bebe805449a41068272703bc92