Extract `Test All Android API levels` into its own class
It is a currently failing (and hence annotated with `@Ignore`) test
that is effectively multiple separate tests. As they are run as one
test only an issue with the first test is reported. Moving it into
its own class allows it to be parameterized so it can run separate
tests making it easier to fix.

Bug: 301282006
Test: ./gradlew
Change-Id: I074a7e1fa26a4b459cfc4609246be05e3c3c97f5